Update naming, folders
